Implements the use of shared memory for keyed perfect hash and baselinehash without reduction on GPU When the output buffer of a query is small enough to fit in shared memory, the query will  utilize a shared m,emory buffer for each block, that will be moved to the corresponding buffer on globacl memory to be copied back t0o system memory for the reduction. This feature can be enabled and disabled using the parameter enable-shared-mem-grouped-all
Given the slowness of the shared memory INT64 atomic sum on RTX class GPUs compared to more expensive GPUs an optimized version AtomicSum64 accelerating  the execution between 3 and 5 times on most of the scenarios. It can be activated using the new parameter enable-shared-mem-optimized-sum.

To give the best performances of RTX hardware, the shared memory path is used depending on the aggregatrions on the query. This behavior can be turned on and off using the parameter enable-shared-mem-weigth.